+++ tests/Unit/CSSList/CSSListTest.php
@@ -1,0 +1,25 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sabberworm\CSS\Tests\Unit\CSSList;
+
+use PHPUnit\Framework\TestCase;
+use Sabberworm\CSS\CSSElement;
+use Sabberworm\CSS\Tests\Unit\CSSList\Fixtures\ConcreteCSSList;
+
+/**
+ * @covers \Sabberworm\CSS\CSSList\CSSList
+ */
+final class CSSListTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function implementsCSSElement(): void
+    {
+        $subject = new ConcreteCSSList();
+
+        self::assertInstanceOf(CSSElement::class, $subject);
+    }
+}